fix: make conversation composer functional [conversation-preview-fixed]
